Code Re factor
