+++ Ginger/Ginger/ALM/Repository/RQMRepository.cs
@@ -241,7 +241,7 @@ namespace Ginger.ALM.Repository
             {
                 return false;
             }
-            var originalExternalFields = GetExternalFields();
+            var originalExternalFields = GingerCoreNET.GeneralLib.General.GetExternalFields();
 
             if (!originalExternalFields.Any(x => x.ItemType == "TestCase"))
             {
@@ -291,45 +291,6 @@ namespace Ginger.ALM.Repository
             return exportRes;
         }
 
-        private static ObservableList<ExternalItemFieldBase> GetExternalFields()
-        {
-            ObservableList<ExternalItemFieldBase> originalExternalFields = new ObservableList<ExternalItemFieldBase>();
-
-            var defaultALMConfig = WorkSpace.Instance.Solution.ALMConfigs.FirstOrDefault(x => x.DefaultAlm);
-            var firstExternalItemField = WorkSpace.Instance.Solution.ExternalItemsFields.FirstOrDefault();
-
-            if (defaultALMConfig != null && firstExternalItemField != null &&
-                defaultALMConfig.ALMProjectGUID != firstExternalItemField.ProjectGuid)
-            {
-                var externalOnlineItemsFields = ImportFromRQM.GetOnlineFields(null);
-
-                foreach (var externalItemField in externalOnlineItemsFields)
-                {
-                    var existingField = WorkSpace.Instance.Solution.ExternalItemsFields
-                        .FirstOrDefault(x => x.Name.Equals(externalItemField.Name, StringComparison.CurrentCultureIgnoreCase));
-
-                    var fieldToAdd = new ExternalItemFieldBase
-                    {
-                        Name = externalItemField.Name,
-                        ID = externalItemField.ID,
-                        ItemType = externalItemField.ItemType,
-                        Guid = externalItemField.Guid,
-                        IsCustomField = externalItemField.IsCustomField,
-                        SelectedValue = existingField != null && !string.IsNullOrEmpty(existingField.SelectedValue)
-                            ? existingField.SelectedValue
-                            : externalItemField.SelectedValue
-                    };
-
-                    originalExternalFields.Add(fieldToAdd);
-                }
-            }
-            else
-            {
-                originalExternalFields = WorkSpace.Instance.Solution.ExternalItemsFields;
-            }
-
-            return originalExternalFields;
-        }
 
         #region External Item Fields
